+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added stream commands to ClusterStrategy
 - Added vector sets commands to ClusterStrategy
 - Added experimental support for `HIMPORT` bulk hash import feature (API may change in a future release)
+- Added `OBJECT` and hash field expiration commands to ClusterStrategy
 
 ### Fixed
 - Fixed Sentinel does not wipe servers on exception caused (#1694)